+++ CHANGELOG.md
@@ -13,9 +13,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - updated as3 class extension to parse/hold as3 declarations for easy access/listing
 - configured atc schema inject/remove function from vscode-f5
-- added file/folder input support so it will support a file or path to load and declare (PENDING)
 - cf class support
   - full info/inspect/declare/trigger/reset endpoint support
+- consolidated ATC models and mocks to export tests for integration with other tools
 
 ---